feat: encapsulation agentHost
